requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
python-telegram-bot==13.15
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.1
tzlocal==5.2
urllib3==2.1.0
